Drop Elasticsearch5SearchResults and make Elasticsearch6SearchResults the base implementation
